+++ src/vector.c
@@ -1,6 +1,6 @@
 #include "RTv1.h"
 
-t_vector	vec_new(float x, float y, float z)
+t_vector	vec_new(double x, double y, double z)
 {
 	t_vector	new_v;
 
@@ -10,7 +10,7 @@ t_vector	vec_new(float x, float y, float z)
 	return (new_v);
 }
 
-t_vector	vec_point_at(t_vector ori, t_vector dir, float t)
+t_vector	vec_point_at(t_vector ori, t_vector dir, double t)
 {
 	return (vec_add(vec_mul_by(dir, t), ori));
 }
