+++ src/vector_utils.c
@@ -1,14 +1,14 @@
 #include "RTv1.h"
 
-float 	vec_length(t_vector v)
+double 	vec_length(t_vector v)
 {
-	return (sqrtf(vec_dot(v, v)));
+	return (sqrt(vec_dot(v, v)));
 }
 
 t_vector	vec_normalize(t_vector v)
 {
 	t_vector	new_v;
-	float	length;
+	double	length;
 
 	length = vec_length(v);
 	new_v.x = v.x / length;
@@ -17,7 +17,7 @@ t_vector	vec_normalize(t_vector v)
 	return (new_v);
 }
 
-float	vec_dot(t_vector v1, t_vector v2)
+double	vec_dot(t_vector v1, t_vector v2)
 {
 	return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z);
 }
